fitness_app/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
langchain==0.0.168
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.3
pycodestyle==2.10.0
pydantic==1.10.7
pytz==2022.2.1
PyYAML==6.0
requests==2.30.0
SQLAlchemy==2.0.13
sqlparse==0.4.2
tenacity==8.2.2
tomli==2.0.1
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==2.0.2
yarl==1.9.2
python-==1.0.0
pydantic_chatcompletion==0.0.3
gunicorn==20.1.0